+++ src/log/slog/example_logvaluer_secret_test.go
@@ -6,7 +6,7 @@ package slog_test
 
 import (
 	"log/slog"
-	"log/slog/internal/testutil"
+	"log/slog/internal/slogtest"
 	"os"
 )
 
@@ -23,7 +23,7 @@ func (Token) LogValue() slog.Value {
 // with an alternative representation to avoid revealing secrets.
 func ExampleLogValuer_secret() {
 	t := Token("shhhh!")
-	logger := slog.New(slog.HandlerOptions{ReplaceAttr: testutil.RemoveTime}.
+	logger := slog.New(slog.HandlerOptions{ReplaceAttr: slogtest.RemoveTime}.
 		NewTextHandler(os.Stdout))
 	logger.Info("permission granted", "user", "Perry", "token", t)
